The file Footer.js has been updated to correct the improper use of class attributes in JSX. The changes made are as follows:
The attribute class has been replaced with className in all JSX tags to adhere to the naming convention for attributes in React.
The attribute fill-rule has been updated to fillRule to fix a syntax error in the SVG definition.
These changes ensure that the code is compatible with React best practices and functions correctly without generating warnings or errors in the browser console.